// web/api/server/models/PackageDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PackageDto = {
    /**
     * The name of the environment
     */
    environment: string;
    /**
     * The name of the package
     */
    name: string;
    /**
     * The URL of the package code (ZIP file)
     */
    codeUrl: string;
};